requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic>=2.5.3
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
jinja2>=3.1.2
greenlet>=3.0.3
